refactor(*): turn monitor into a package
