+++ src/command/__init__.py
@@ -1,3 +1,3 @@
 from __future__ import annotations
 
-from . import opml, sub, utils, monitor, customization, administration, misc
+from . import opml, sub, utils, customization, administration, misc
