+++ src/monitor/__init__.py
@@ -1,0 +1,1 @@
+from ._monitor import Monitor
